Improved checks and added logout function
Checks now use a second file with the .time extension. It stores the expiration time of each session. Whenever a check is done for an expired session, the session files are deleted.

I also added a logout function that simple deletes the session files on the server, and the marks the session cookie as expired.